+++ lib/application/user/state/create_user_result.dart
@@ -1,9 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../domain/repository/user/entity/user.dart';
-
 /// 匿名ユーザーを作成結果プロバイダー
-final createUserResultProvider = StateProvider<AsyncValue<User?>>(
+final createUserResultProvider = StateProvider<AsyncValue<void>>(
   (_) => const AsyncValue.data(null),
   name: 'createUserResultProvider',
 );
