+++ lib/application/user/user_service.dart
@@ -42,7 +42,6 @@ class UserService {
           createdPlatform: AppPlatform.currentPlatform,
         ),
       );
-      return user;
     });
   }
